Added new PNDelegate method which will be pulled to check whether application is able to run in background and thus PubNub client should or not - (BOOL)shouldRunClientInBackgournd;
WARNING: Don't set it to 'YES' if you extending applications' background lifetime with beginBackgroundTaskWithExpirationHandler: and only if application can
persistently run in background. You can not implement this method and specify corresponding background modes in applications' PLIST and library will read out from it (but careful with location, because if manager doesn't run, applicatino willbe suspended)

In case if application doesn't support background execution context, connections will be suspended and on application wake will be resumed.